+++ CHANGELOG.md
@@ -5,6 +5,8 @@ We follow [Semantic Versions](https://semver.org/).
 ## Unreleased
 
 - Improve `django_adjust_s3_bucket` to take in account different storages
+- Make `_validate_file_existence` use specified storage
+- Add storage setting for `S3UploadURLField`
 
 ## 0.4.0