+++ CHANGELOG.md
@@ -4,6 +4,10 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Unreleased
 
+- Improve `django_adjust_s3_bucket` to take in account different storages
+- Make `_validate_file_existence` use specified storage
+- Add storage setting for `S3UploadURLField`
+
 ## 0.4.0
 
 - Extend api to list/retrieve configs for s3 uploads in django module